Fix editor not playing audio due to video order
If the editor has multiple videos, it will consider
one of them the "primary" video, from which it will
play back audio. Even if that "primary" video does
not have any audio.
This changes the logic so that the first video
with audio the editor comes across is made the
"primary" video, so that there should always
be audio playing if audio is available.